requirements.txt

aiosqlite==0.16.1
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.23.0
certifi==2021.5.30
chardet
click==7.1.2
ddlparse==1.9.0
dictdiffer==0.8.1
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.2
FastAPI-SQLAlchemy==0.2.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
httptools==0.1.2
idna==2.10
install==1.3.4
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
orjson==3.5.3
promise==2.3
pydantic==1.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-multipart==0.0.5
pytz
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.18
starlette==0.14.2

typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1

pyjwt
passlib==1.7.2
bcrypt==3.1.7
psycopg2-binary==2.8.5
Authlib==0.14.3
httpx==0.14.0
typer==0.3.2
tortoise-orm
aerich
pytest==6.1.2
alembic==1.6.5
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
certifi==2021.5.30
click
colorama==0.4.4
distlib==0.3.2
filelock==3.0.12
greenlet==1.1.0
h11
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
pipenv==2021.5.29
psycopg2==2.9.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
regex==2021.4.4
aiofiles==0.7.0
toml==0.10.2
uvicorn
virtualenv-clone==0.5.4
msgpack-asgi==1.0.0
reportlab==3.5.68
tortoise-orm[asyncpg]==0.16.19
tortoise-cli==0.1.1
emails==0.6
mongoengine==0.23.1
pymongo==3.12.0
motor==2.5.0